+++ packages/angular/src/lib/animations/animations.module.ts
@@ -1,10 +1,8 @@
-import { NgModule, Injectable, Inject, NgZone, RendererFactory2, Optional, SkipSelf } from '@angular/core';
+import { NgModule, Injectable, Inject, NgZone, RendererFactory2, Optional, SkipSelf, ɵChangeDetectionScheduler as ChangeDetectionScheduler, inject } from '@angular/core';
 import { DOCUMENT } from '@angular/common';
-import { AnimationBuilder } from '@angular/animations';
+import { AnimationBuilder, ɵBrowserAnimationBuilder as BrowserAnimationBuilder } from '@angular/animations';
 
-import { AnimationDriver, ɵAnimationStyleNormalizer as AnimationStyleNormalizer, ɵWebAnimationsStyleNormalizer as WebAnimationsStyleNormalizer, ɵAnimationEngine as AnimationEngine } from '@angular/animations/browser';
-
-import { ɵAnimationRendererFactory as AnimationRendererFactory, ɵBrowserAnimationBuilder as BrowserAnimationBuilder } from '@angular/platform-browser/animations';
+import { AnimationDriver, ɵAnimationRendererFactory as AnimationRendererFactory, ɵAnimationStyleNormalizer as AnimationStyleNormalizer, ɵWebAnimationsStyleNormalizer as WebAnimationsStyleNormalizer, ɵAnimationEngine as AnimationEngine } from '@angular/animations/browser';
 
 // import { NativeScriptModule } from "../nativescript.module";
 import { NativeScriptRendererFactory } from '../nativescript-renderer';
@@ -15,7 +13,7 @@ import { NativeScriptCommonModule } from '../nativescript-common.module';
 @Injectable()
 export class InjectableAnimationEngine extends AnimationEngine {
   constructor(@Inject(DOCUMENT) doc: any, driver: AnimationDriver, normalizer: AnimationStyleNormalizer) {
-    super(doc.body, driver, normalizer);
+    super(doc, driver, normalizer);
   }
 }